Working on uploading images, not currently working

+++ app/views/posts/show.blade.php
@@ -1,16 +1,23 @@
 @extends("layouts.master")
 
 @section('content')
-	<p>Created on {{ $post->created_at->format('l, F jS Y @ h:i:s A') }}
-	<h1>{{{$post->title}}}</h1>
-	<h2>{{{$post->body}}}</h2>
-	@if(Auth::user()->username  == $post->user->username)
-		<a href="{{{ action('PostsController@edit' , $post->id)}}}">Edit post</a>
-		{{ Form::open(array('action' =>array('PostsController@destroy', $post->id), 'method' => 'DELETE' , 'id' => 'formDelete')) }}
-			<button id="deleteButton">Delete post</button>
-		{{ Form::close() }}
-	@endif
-	
+	<div class="showDiv">
+		<p class="glyphicon glyphicon-time">Created {{ $post->created_at->format('l, F jS Y @ h:i:s A') }}
+		<h1>{{{$post->title}}}</h1>
+		<h2>{{{$post->body}}}</h2>
+		@if(Auth::check())
+			@if(Auth::user()->username  == $post->user->username)
+				<a href="{{{ action('PostsController@edit' , $post->id)}}}">Edit post</a>
+				{{ Form::open(array('action' =>array('PostsController@destroy', $post->id), 'method' => 'DELETE' , 'id' => 'formDelete')) }}
+					<button id="deleteButton">Delete post</button>
+				{{ Form::close() }}
+			@endif
+		@endif
+		@if(Storage::exists("$post->picture"))
+   			<img src="/files/{{ "$post->picture"}}" alt="" height="214" width="500">
+		@endif
+	</div>
+
 @stop
 
 @section('script')
